LexerTests: refactor to use Verify
