+++ Cesium.Parser.Tests/LexerTests/LexerTests.SimpleTest.verified.txt
@@ -1,0 +1,6 @@
+﻿KeywordInt: int
+Identifier: main
+OpenParen: (
+CloseParen: )
+OpenBrace: {
+CloseBrace: }
